+++ dataservices/management/commands/import_easeofdoingbusiness_data.py
@@ -1,5 +1,6 @@
 import tablib
 from django.core.management import BaseCommand
+from django.db import connection
 from import_export import resources
 
 from dataservices.models import EaseOfDoingBusiness
@@ -12,10 +13,13 @@ class Command(BaseCommand):
         with open('dataservices/resources/EaseOfDoingBusiness.csv', 'r', encoding='utf-8-sig') as f:
             data = tablib.import_set(f.read(), format='csv', headers=True)
             easeofdoingbusiness_resource = resources.modelresource_factory(model=EaseOfDoingBusiness)()
-            result = easeofdoingbusiness_resource.import_data(data, dry_run=True)
-            self.stdout.write(self.style.SUCCESS(result.has_errors()))
-            if not result.has_errors():
-                # No Errors lets flush table and import the data
-                EaseOfDoingBusiness.objects.all().delete()
-                easeofdoingbusiness_resource.import_data(data, dry_run=False)
+            EaseOfDoingBusiness.objects.all().delete()
+            easeofdoingbusiness_resource.import_data(data, dry_run=False)
+            self.stdout.write('Linking countries')
+            cursor = connection.cursor()
+            cursor.execute(
+                "update dataservices_easeofdoingbusiness as d \
+                set country_id=c.id \
+                from dataservices_country c where d.country_code=c.iso3;"
+            )
         self.stdout.write(self.style.SUCCESS('All done, bye!'))